+++ sdk/healthbot/arm-healthbot/src/operationsInterfaces/operations.ts
@@ -7,7 +7,10 @@
  */
 
 import { PagedAsyncIterableIterator } from "@azure/core-paging";
-import { OperationDetail, OperationsListOptionalParams } from "../models/index.js";
+import {
+  OperationDetail,
+  OperationsListOptionalParams,
+} from "../models/index.js";
 
 /// <reference lib="esnext.asynciterable" />
 /** Interface representing a Operations. */
@@ -17,6 +20,6 @@ export interface Operations {
    * @param options The options parameters.
    */
   list(
-    options?: OperationsListOptionalParams
+    options?: OperationsListOptionalParams,
   ): PagedAsyncIterableIterator<OperationDetail>;
 }