+++ sdk/healthbot/arm-healthbot/CHANGELOG.md
@@ -1,15 +1,33 @@
 # Release History
 
-## 2.1.1 (Unreleased)
+## 3.0.0 (2025-08-19)
 
 ### Features Added
+  - Added operation Bots.beginUpdate
+  - Added operation Bots.beginUpdateAndWait
+  - Added operation Bots.listSecrets
+  - Added operation Bots.regenerateApiJwtSecret
+  - Added Interface BotsListSecretsOptionalParams
+  - Added Interface BotsRegenerateApiJwtSecretOptionalParams
+  - Added Interface HealthBotKey
+  - Added Interface HealthBotKeysResponse
+  - Added Interface KeyVaultProperties
+  - Interface BotsUpdateOptionalParams has a new optional parameter resumeFrom
+  - Interface BotsUpdateOptionalParams has a new optional parameter updateIntervalInMs
+  - Interface HealthBotProperties has a new optional parameter accessControlMethod
+  - Interface HealthBotProperties has a new optional parameter keyVaultProperties
+  - Interface HealthBotUpdateParameters has a new optional parameter properties
+  - Added Type Alias BotsListSecretsResponse
+  - Added Type Alias BotsRegenerateApiJwtSecretResponse
 
 ### Breaking Changes
+  - Removed operation Bots.update
+  - Operation Bots.beginCreate has a new signature
+  - Operation Bots.beginCreateAndWait has a new signature
+  - Operation Bots.get has a new signature
+  - Type alias "SkuName" has been changed
 
-### Bugs Fixed
-
-### Other Changes
-
+    
 ## 2.1.0 (2022-11-25)
     
 ### Features Added
@@ -26,4 +44,4 @@ To understand the detail of the change, please refer to [Changelog](https://aka.
 
 To migrate the existing applications to the latest version, please refer to [Migration Guide](https://aka.ms/js-track2-migration-guide).
 
-To learn more, please refer to our documentation [Quick Start](https://aka.ms/azsdk/js/mgmt/quickstart ).
+To learn more, please refer to our documentation [Quick Start](https://aka.ms/azsdk/js/mgmt/quickstart).